+++ src/dictionary/dicLoader/dicLoader.h
@@ -16,7 +16,7 @@ typedef struct _dicLoader
 DICLOADER *     dicLoader_new(const char *filename);
 void            dicLoader_free(DICLOADER *dicLoader);
 
-uint32_t        dicLoader_getQtdInst(DICLOADER *dicLoader);
+uint64_t        dicLoader_getQtdInst(DICLOADER *dicLoader);
 uint32_t        dicLoader_getNextInst(DICLOADER *dicLoader);
 
 #endif /* DICLOADER_HEADER */
